requirements.txt
# ========================================
# Frontend UI - Core Dependencies
# ========================================

# Gradio UI Framework
gradio==4.44.0

# HTTP Client (for calling backend API)
httpx==0.27.2
requests==2.32.3

# Text Processing (for file uploads)
python-docx==1.0.1
docx2txt==0.8
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0

# Data Processing (for displaying results)
pandas==2.0.3

# Utilities
Pillow==10.4.0
langid==1.1.6

# Environment & Config
python-dotenv==1.0.1

# Note: No ML libraries needed in frontend!
# Backend handles all ML processing via API
